Exception Handling Implemented

+++ src/main/java/org/scaler/universitysystem/service/ApplicantServiceImpl.java
@@ -1,6 +1,8 @@
 package org.scaler.universitysystem.service;
 
 
+import org.scaler.universitysystem.exceptions.AdmissionNotFoundException;
+import org.scaler.universitysystem.exceptions.ApplicantNotFoundException;
 import org.scaler.universitysystem.models.Applicant;
 import org.scaler.universitysystem.repository.ApplicantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,11 +27,19 @@ public class ApplicantServiceImpl implements ApplicantService{
 
     @Override
     public Optional<Applicant> getApplicantById(Long id) {
+        Optional<Applicant> applicant = applicantRepository.findById(id);
+        if (applicant.isEmpty()){
+            throw new ApplicantNotFoundException(id);
+        }
         return applicantRepository.findById(id);
     }
 
     @Override
     public Applicant saveApplicant(Applicant applicant) {
+        //if name or date or email or dateOfBirth is null, throw an exception
+        if(applicant.getName() == null || applicant.getEmail() == null || applicant.getAddress() == null || applicant.getDateOfBirth() == null){
+            throw new IllegalArgumentException("Name, Email, Address and Date of Birth are mandatory fields");
+        }
         return applicantRepository.save(applicant);
     }
 
@@ -41,12 +51,16 @@ public class ApplicantServiceImpl implements ApplicantService{
             return applicantRepository.save(update(existingApplicant.get(), applicant));
         } else {
             // Throw an error or return null
-            throw new RuntimeException("Applicant not found with id: " + applicant.getId());
+            throw new ApplicantNotFoundException(applicant.getId());
         }
     }
 
     @Override
     public void deleteApplicantById(Long id) {
+        Optional<Applicant> applicant = applicantRepository.findById(id);
+        if (applicant.isEmpty()){
+            throw new ApplicantNotFoundException(id);
+        }
         applicantRepository.deleteById(id);
     }
