+++ src/main/java/org/scaler/universitysystem/service/SelfProgramService.java
@@ -1,5 +1,7 @@
 package org.scaler.universitysystem.service;
 
+import org.scaler.universitysystem.exceptions.DepartmentNotFoundException;
+import org.scaler.universitysystem.exceptions.ProgramNotFoundException;
 import org.scaler.universitysystem.models.Department;
 import org.scaler.universitysystem.models.Program;
 import org.scaler.universitysystem.repository.DepartmentRepository;
@@ -25,13 +27,29 @@ public class SelfProgramService implements ProgramService {
 
     @Override
     public Program getProgramById(Long id) {
+
         Optional<Program> optionalProgram = programRepository.findById(id);
+        if (optionalProgram.isEmpty()) {
+            throw new ProgramNotFoundException(id);
+        }
         return optionalProgram.get();
     }
 
     @Override
     public Program createProgram(Program program) {
+        //if name or duration or degreeLevel or department is null, throw an exception
+        if (program.getName() == null || program.getDuration() == 0 || program.getDegreeLevel() == null || program.getDepartment() == null) {
+            throw new IllegalArgumentException("Name, Duration, Degree Level and Department are mandatory fields");
+        }
+
+
         Department department = program.getDepartment();
+        if (department.getId()!=null){
+            Optional<Department> optionalDepartment = departmentRepository.findById(department.getId());
+            if (optionalDepartment.isEmpty()) {
+                throw new DepartmentNotFoundException(department.getId());
+            }
+        }
         if (department.getId() == null) {
            department = departmentRepository.save(department);
          //   department = departmentService.createDepartment(department);
@@ -51,18 +69,31 @@ public class SelfProgramService implements ProgramService {
         if (optionalProgram.isPresent()) {
             return programRepository.save(update(optionalProgram.get(), program));
         } else {
-            throw new RuntimeException("Program not found with id: " + id);
+            throw new ProgramNotFoundException(id);
         }
     }
 
     @Override
     public void deleteProgram(Long id) {
+        Optional<Program> program = programRepository.findById(id);
+        if (program.isEmpty()) {
+            throw new ProgramNotFoundException(id);
+        }
         programRepository.deleteById(id);
     }
 
     @Override
     public List<Program> getProgramsByDepartment(Long departmentId) {
+        Optional<Department> department = departmentRepository.findById(departmentId);
+        if (department.isEmpty()) {
+            throw new DepartmentNotFoundException(departmentId);
+        }
         return programRepository.findByDepartmentId(departmentId);
+    }
+
+    @Override
+    public List<Program> getAllPrograms() {
+        return programRepository.findAll();
     }
 
     public Program update(Program existingProgram, Program updatedProgram) {
